moved optimizablesymmetrystorage to have a cpp impl
