+++ source/rigidbody/CMakeLists.txt
@@ -29,6 +29,7 @@ add_library(ausaxs_rigidbody OBJECT
 	
 	"parameters/BodyTransformParametersAbsolute.cpp"
 	"parameters/BodyTransformParametersRelative.cpp"
+	"parameters/OptimizableSymmetryStorage.cpp"
 	"parameters/ParameterGenerationFactory.cpp"
 	"parameters/ParameterGenerationStrategy.cpp"
 	"parameters/ParameterGenerationStrategies.cpp"
